+++ jest.config.js
@@ -1,14 +1,22 @@
 // jest.config.js
 
 // This Jest configuration is designed for a TypeScript project using ESM modules with ts-jest.
+// @ts-check
 
 import { createDefaultEsmPreset } from 'ts-jest';
 
 // Create an ESM configuration to process TypeScript files (.ts/.mts/.tsx/.mtsx).
-const presetConfig = createDefaultEsmPreset({
-  tsconfig: './tsconfig.jest.json',
-});
+/** @typedef {{ tsconfig: string }} TsJestEsmPresetOptions */
 
+/** @type {TsJestEsmPresetOptions} */
+const tsJestEsmPresetOptions = {
+  tsconfig: './tsconfig.jest.json',
+};
+
+/** @type {import('ts-jest').DefaultEsmPreset} */
+const presetConfig = createDefaultEsmPreset(tsJestEsmPresetOptions);
+
+/** @type {import('ts-jest').JestConfigWithTsJest} */
 const jestConfig = {
   ...presetConfig,
   testEnvironment: 'node', // Use Node.js environment for testing