+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Advantages:
 - [MdnsReflectorServer]: Added --share-with-clients to share mDNS between all reflector clients.
 - [reflector]: [Full guide](https://github.com/Luligu/matterbridge/blob/dev/docker-reflector/Docker%20reflector.md).
 - [jest]: Refactored configuration with cross-env.
+- [jest]: Typed jest config.
 - [eslint]: Refactored configuration.
 
 ### Changed